Add CNNIC SSL to blacklist

+++ Android/flashable.sh
@@ -5,7 +5,7 @@ GEN_ZIP='flashable.zip'
 
 mkdir -p flashable/data/misc/keychain
 
-rm $GEN_ZIP
+test -f $GEN_ZIP && rm $GEN_ZIP
 
 cp pubkey_blacklist.txt flashable/data/misc/keychain
 cp serial_blacklist.txt flashable/data/misc/keychain
